+++ src/uwtools/tests/drivers/test_support.py
@@ -55,8 +55,9 @@ def test_tasks():
 
     assert support.tasks(SomeDriver) == {
         "run": "A run.",
+        "runscript": "The runscript.",
+        "t1": "@external t1",
         "t2": "@task t2",
         "t3": "@tasks t3",
-        "t1": "@external t1",
         "validate": "Validate the UW driver config.",
     }
